Move the shutdown handler tests into their own file
The six tests share no infrastructure with the repeater tests,
so they live in tests/test_shutdown.py, matching the one-file-per-topic
convention and keeping test_repeater.py about the Repeater.
